refactor: move class for re-package structure
SCRUM-24
