+++ src/main/java/com/codecrafters/companity/comment/adapter/dto/response/ResponseComment.java
@@ -1,11 +1,11 @@
-package com.codecrafters.companity.adapter.comment.dto.response;
+package com.codecrafters.companity.comment.adapter.dto.response;
 
-import com.codecrafters.companity.domain.comment.Comment;
+import com.codecrafters.companity.comment.domain.Comment;
 import com.codecrafters.companity.domain.user.User;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import static com.codecrafters.companity.adapter.comment.CommentMapper.COMMENT_MAPPER;
+import static com.codecrafters.companity.comment.adapter.CommentMapper.COMMENT_MAPPER;
 
 import java.time.LocalDateTime;
