+++ src/main/java/com/codecrafters/companity/comment/adapter/infrastructure/jpa/CommentEntity.java
@@ -1,4 +1,4 @@
-package com.codecrafters.companity.adapter.comment.infrastructure.jpa;
+package com.codecrafters.companity.comment.adapter.infrastructure.jpa;
 
 import com.codecrafters.companity.adapter.common.infrastructure.jpa.BaseTimeEntity;
 import com.codecrafters.companity.adapter.post.infrastructure.jpa.PostEntity;
