+++ src/main/java/com/codecrafters/companity/comment/application/port/CommentService.java
@@ -1,9 +1,9 @@
-package com.codecrafters.companity.application.service.comment;
+package com.codecrafters.companity.comment.application.port;
 
-import com.codecrafters.companity.application.in.comment.CommentUseCase;
-import com.codecrafters.companity.application.out.persistence.CommentRepository;
-import com.codecrafters.companity.domain.comment.CommentForCreate;
-import com.codecrafters.companity.domain.comment.Comment;
+import com.codecrafters.companity.comment.application.port.in.CommentUseCase;
+import com.codecrafters.companity.comment.application.port.out.persistence.CommentRepository;
+import com.codecrafters.companity.comment.domain.CommentForCreate;
+import com.codecrafters.companity.comment.domain.Comment;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
